# contracts/catalog/openapi.json
{
  "components": {
    "schemas": {
...
      "ActionRequest": {
        "description": "Request model for executing a role action.",
        "properties": {
          "command": {
            "title": "Command",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "default": {},
            "title": "Params",
            "type": "object"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "command"
        ],
        "title": "ActionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScenarioResponse": {
        "description": "Base response model for scenario operations.\n\nManual notes from the activation (e.g. \"set the Dodocus to LD\") are NOT on this\nresponse — they live on ``ScenarioState.manual_steps`` (single source of truth, fetched\nvia ``GET /scenario/state``; survives page reload).",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "ScenarioResponse",
        "type": "object"
      },
...
      "ShutdownScenarioRequest": {
        "description": "Request model for shutting down a scenario.",
        "properties": {
          "graceful": {
            "default": true,
            "title": "Graceful",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "ShutdownScenarioRequest",
        "type": "object"
      },
      "StartScenarioRequest": {
        "description": "Request model for starting a scenario.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "StartScenarioRequest",
        "type": "object"
      },
...
      "SwitchScenarioRequest": {
        "description": "Request model for switching scenarios.",
        "properties": {
          "graceful": {
            "default": true,
            "title": "Graceful",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "SwitchScenarioRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "A web service that manages MQTT devices with typed configurations",
    "title": "MQTT Web Service",
    "version": "0.6.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/scenario/definition": {
      "get": {
        "description": "Get definitions of scenarios, optionally filtered by room.\n\nArgs:\n    room: Optional room ID to filter scenarios by\n    \nReturns:\n    List[ScenarioDefinition]: List of scenario definitions\n    \nRaises:\n    HTTPException: If service not initialized",
        "operationId": "get_scenarios_for_room_scenario_definition_get",
        "parameters": [
          {
            "description": "Filter scenarios by room ID",
            "in": "query",
            "name": "room",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter scenarios by room ID",
              "title": "Room"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScenarioDefinition"
                  },
                  "title": "Response Get Scenarios For Room Scenario Definition Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scenarios For Room",
        "tags": [
          "Scenarios"
        ]
      }
    },
    "/scenario/definition/{id}": {
      "get": {
        "description": "Get the definition of a specific scenario.\n\nReturns:\n    ScenarioDefinition: The scenario definition\n    \nRaises:\n    HTTPException: If scenario not found or service not initialized",
        "operationId": "get_scenario_definition_scenario_definition__id__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDefinition"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scenario Definition",
        "tags": [
          "Scenarios"
        ]
      }
    },
    "/scenario/role_action": {
      "post": {
        "description": "Execute an action on a device bound to a role in the current scenario.\n\nArgs:\n    data: The role action request with role, command and params\n    \nReturns:\n    Dict[str, Any]: The result of the command execution\n    \nRaises:\n    HTTPException: If no active scenario or an error occurs",
        "operationId": "execute_role_action_scenario_role_action_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Execute Role Action Scenario Role Action Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Role Action",
        "tags": [
          "Scenarios"
        ]
      }
    },
    "/scenario/shutdown": {
      "post": {
        "description": "Shutdown the currently active scenario.\n\nThis endpoint shuts down the currently active scenario if it matches the provided ID.\nIf no scenario is active or the active scenario doesn't match the provided ID, \nit returns an appropriate error.\n\nArgs:\n    data: The shutdown scenario request with scenario ID and graceful flag\n    \nReturns:\n    ScenarioResponse: Status of the operation\n    \nRaises:\n    HTTPException: If no active scenario, scenario mismatch, or an error occurs",
        "operationId": "shutdown_scenario_scenario_shutdown_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShutdownScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Shutdown Scenario",
        "tags": [
          "Scenarios"
        ]
      }
    },
    "/scenario/start": {
      "post": {
        "description": "Start a scenario if no scenario is currently active.\n\nThis endpoint starts a scenario only if there is no currently active scenario.\nIf a scenario is already running, it returns an error.\n\nArgs:\n    data: The start scenario request with scenario ID\n    \nReturns:\n    ScenarioResponse: Status of the operation\n    \nRaises:\n    HTTPException: If scenario not found, another scenario is active, or an error occurs",
        "operationId": "start_scenario_scenario_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Scenario",
        "tags": [
          "Scenarios"
        ]
      }
    },
...
    "/scenario/switch": {
      "post": {
        "description": "Switch to a different scenario.\n\nThis endpoint performs a transition between scenarios, handling\ndevice state changes efficiently.\n\nArgs:\n    data: The switch scenario request with scenario ID and graceful flag\n    \nReturns:\n    ScenarioResponse: Status of the operation\n    \nRaises:\n    HTTPException: If scenario not found or an error occurs",
        "operationId": "switch_scenario_scenario_switch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Switch Scenario",
        "tags": [
          "Scenarios"
        ]
      }
    },
    "/scenario/{id}/force_reconcile": {
      "post": {
        "description": "Force ONE device into the active scenario's desired state — the\nbelieved-vs-desired diff is skipped (the belief may be wrong; the user picking the\nrow is the feedback channel), driver idempotence guards are bypassed via the\nreserved `force` param, and toggle power claims the plan target (`assume_state`).\nRuns the device's chain through the normal executor (gates + polls); worst case a\npoll-timeout wait, so the call can take seconds.",
        "operationId": "force_reconcile_device_scenario__id__force_reconcile_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForceReconcileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForceReconcileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Force Reconcile Device",
        "tags": [
          "Scenarios"
        ]
      }
    },
    "/scenario/{id}/layout": {
      "get": {
        "description": "Layer-3 layout manifest for a scenario — the composite remote (one renderer;\ncontrols carry their canonical capability/action and dispatch against the manifest's\n`canonicalEntityId`; the power zone is the scenario lifecycle). Built from the\nscenario definition + the role devices' capability maps by the placement engine. The\n`inputs` role is intentionally not rendered (reconciler-derived).",
        "operationId": "get_scenario_layout_scenario__id__layout_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayoutManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scenario Layout",
        "tags": [
          "Scenarios"
        ]
      }
    },
    "/scenario/{id}/reconcile_preview": {
      "get": {
        "description": "Believed-vs-desired state per involved device of the ACTIVE scenario, plus the\nforced chain a confirm would run. 404 unknown scenario; 409 when it isn't the active\none (the desired state is only defined by the running scenario — on an inactive page\nthe same gesture is just \"start it\").",
        "operationId": "get_reconcile_preview_scenario__id__reconcile_preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconcilePreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Reconcile Preview",
        "tags": [
          "Scenarios"
        ]
      }
    },
...
  }
}